+++ plugins/tests/adapters/braintrust/__init__.py
@@ -1,0 +1,1 @@
+"""Braintrust adapter tests."""